tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/blacklisted-service-in-filter.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';
test(function() {
    window.testRunner.setBluetoothMockDataSet('EmptyAdapter');
    assert_throws('SecurityError', function () {
        window.navigator.bluetooth.requestDevice({filters: [{services: ['human_interface_device']}]});
    });
}, 'Reject with SecurityError if requesting a blacklisted service.');
</script>
